Add rootReducer to combine all reducers and refactor code
