+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/aldana/Microverse/bookstore/src/index.js":"1","/home/aldana/Microverse/bookstore/src/containers/BooksList.js":"2","/home/aldana/Microverse/bookstore/src/components/BookForm.js":"3","/home/aldana/Microverse/bookstore/src/redux/reducers/books.js":"4"},{"size":360,"mtime":1611588095186,"results":"5","hashOfConfig":"6"},{"size":667,"mtime":1611588071934,"results":"7","hashOfConfig":"6"},{"size":671,"mtime":1611587890692,"results":"8","hashOfConfig":"6"},{"size":356,"mtime":1611585427907,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ipiu3q",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/aldana/Microverse/bookstore/src/index.js",[],"/home/aldana/Microverse/bookstore/src/containers/BooksList.js",[],"/home/aldana/Microverse/bookstore/src/components/BookForm.js",[],"/home/aldana/Microverse/bookstore/src/redux/reducers/books.js",[]]
+[{"/home/aldana/Microverse/bookstore/src/redux/reducers/books.js":"1"},{"size":760,"mtime":1611646488519,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ipiu3q","/home/aldana/Microverse/bookstore/src/redux/reducers/books.js",[]]
