+++ src/redux/reducers/books.js
@@ -1,7 +1,15 @@
 const CREATE_BOOK = 'CREATE_BOOK';
 const REMOVE_BOOK = 'REMOVE_BOOK';
 
-const booksReducer = (state, action) => {
+const initialState = [
+  { id: Math.random(), title: 'Title for book #1', category: 'Horror' },
+  { id: Math.random(), title: 'Title for book #2', category: 'Kids' },
+  { id: Math.random(), title: 'Title for book #3', category: 'History' },
+  { id: Math.random(), title: 'Title for book #4', category: 'Horror' },
+  { id: Math.random(), title: 'Title for book #5', category: 'Kids' },
+];
+
+const booksReducer = (state = initialState, action) => {
   switch (action.type) {
     case CREATE_BOOK:
       return [...state, action.payload];
