+++ src/redux/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import booksReducer from './books';
+
+const rootReducer = combineReducers({ booksReducer });
+
+export default rootReducer;
